refactor(client): rename AssignMinersModal → ManageRackModal (#395)
Align the rack modal naming with `ManageBuildingModal` and `ManageSiteModal`.
Renames the directory, default-exported component, and props interface.
Updates importers and historical comments that reference the old name.
