--- client/src/protoFleet/features/fleetManagement/components/AssignMinersModal/index.ts
@@ -1,2 +1,0 @@
-export { default as AssignMinersModal } from "./AssignMinersModal";
-export type { RackFormData, AssignmentMode } from "./types";
