+++ client/src/protoFleet/features/fleetManagement/components/ManageRackModal/MinersPane.tsx
