+++ client/src/protoFleet/features/fleetManagement/components/ManageRackModal/index.ts
@@ -1,0 +1,2 @@
+export { default as ManageRackModal } from "./ManageRackModal";
+export type { RackFormData, AssignmentMode } from "./types";
